refactor for consistent coding style
WARNING: Untested code ahead, DO NOT MERGE INTO MAIN
